feat: convert to TypeScript
